+++ duckscript_sdk/src/sdk/std/net/ftp/put_in_memory/mod_test.rs
@@ -1,0 +1,12 @@
+use super::*;
+use crate::test;
+
+#[test]
+fn common_functions() {
+    test::test_common_command_functions(create(""));
+}
+
+#[test]
+fn run_no_args() {
+    test::run_script_and_error(vec![create("")], "out = ftp_put_in_memory", "out");
+}