+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## CHANGELOG
 
-### v0.5.1
+### v0.6.0
 
 * New ftp_put command.
 * New fpt_put_in_memory command.
@@ -8,6 +8,7 @@
 * New ftp_get_in_memory command.
 * New ftp_list command.
 * New ftp_nlst command.
+* Conditions to support function calls #116
 
 ### v0.5.0 (2020-06-06)